move BackgroundDimmer inside Modal, do not re-export it
