--- src/BackgroundDimmer/examples.scss
@@ -1,32 +1,0 @@
-@import '../utils/theme/variables.scss';
-
-.background-dimmer {
-  .card {
-    padding: 38px;
-    border-radius: 4px;
-    background-color: $brc-white;
-    color: $brc-darkGrey;
-    font-size: 16px;
-    line-height: 1.2;
-
-    > img {
-      margin-bottom: 14px;
-    }
-
-    .card-title {
-      color: $brc-darkGrey;
-      font-size: 16px;
-      font-weight: $brc-semibold;
-    }
-
-    .card-author {
-      margin-bottom: 14px;
-      color: $brc-coolGrey;
-      font-size: 14px;
-    }
-
-    .card-rating {
-      height: 12px;
-    }
-  }
-}
