--- src/BackgroundDimmer/index.ts
@@ -1,2 +1,0 @@
-import { BackgroundDimmer } from './BackgroundDimmer';
-export default BackgroundDimmer;
